agetty: split out terminal functions to tty.c
Move terminal initialization and management functions to tty.c:
agetty_open_tty(), agetty_termio_clear(), agetty_termio_init(),
agetty_termio_final(), agetty_reset_vc(), agetty_auto_baud(),
agetty_next_speed() and agetty_erase_char().

Move CLOCAL_MODE_* enum, FIRST_SPEED and debug() macro to agetty.h
as they are now needed by multiple source files.
